削除機能

+++ src/main/java/scoremanager/main/SubjectDeleteExecuteAction.java
@@ -1,5 +1,3 @@
-// SubjectDeleteExecuteAction.java
-
 package scoremanager.main;
 
 import java.io.IOException;
@@ -14,14 +12,17 @@ import tool.Action;
 
 public class SubjectDeleteExecuteAction extends Action {
 
-    public void execute(HttpServletRequest req, HttpServletResponse res)
+    public void execute(
+            HttpServletRequest req,
+            HttpServletResponse res)
             throws Exception, IOException {
 
-        // ログイン情報
-        Teacher teacher = (Teacher) req.getSession()
-                .getAttribute("user");
+        // ログイン情報取得
+        Teacher teacher =
+            (Teacher) req.getSession()
+            .getAttribute("user");
 
-        // 学校取得
+        // 学校情報取得
         School school = teacher.getSchool();
 
         // 科目コード取得
@@ -29,13 +30,36 @@ public class SubjectDeleteExecuteAction extends Action {
 
         // DAO生成
         SubjectDao dao = new SubjectDao();
+
+        // 科目取得
         Subject subject = dao.get(cd, school);
 
+        // 科目が存在しない場合
+        if (subject == null) {
+
+            req.getRequestDispatcher(
+                "/scoremanager/main/subject_list.jsp"
+            ).forward(req, res);
+
+            return;
+        }
+
         // 削除実行
-        dao.delete(subject); // TODO: 失敗(==false)でerror.jspにリダイレクト
+        boolean result = dao.delete(subject);
 
-        // 一覧画面へ戻る
-        res.sendRedirect("SubjectList.action"); // TODO: SubjectDeleteDone.actionにリダイレクト
+        // 削除失敗
+        if (!result) {
+
+            req.getRequestDispatcher(
+                "/error.jsp"
+            ).forward(req, res);
+
+            return;
+        }
+
+        // 完了画面へ
+        req.getRequestDispatcher(
+            "/scoremanager/main/subject_delete_done.jsp"
+        ).forward(req, res);
     }
-
 }
